Bug Fix: Frame Size Should Be 64-bits
wsServer assumed that 'size_t' always had a size of 64 bits, which is
wrong and does not work on 32-bit architectures. This commit corrects
this by using 'uint64_t' in all frame size references.

Thanks to @Michanord for reporting this bug.
Fixes #14.

+++ example/send_receive.c
@@ -73,12 +73,12 @@ void onclose(int fd)
  *
  * @param type Message type.
  */
-void onmessage(int fd, const unsigned char *msg, size_t size, int type)
+void onmessage(int fd, const unsigned char *msg, uint64_t size, int type)
 {
 	char *cli;
 	cli = ws_getaddress(fd);
-	printf("I receive a message: %s (size: %zu, type: %d), from: %s/%d\n", msg, size,
-		type, cli, fd);
+	printf("I receive a message: %s (size: %" PRId64 ", type: %d), from: %s/%d\n",
+		msg, size, type, cli, fd);
 	free(cli);
 
 	/**
